Added first batch files from my JupyterLab project
I could not load all of the image files for training and testing the CNN due to size and quantity restrictions so I settled for just loading the images from verification to allow small demonstrations on the final data product, of course any .jpg. or .jpeg file could be used.   The .h5 trained model save file is too large for browser upload, so I will try command line method.